==> 5-1_Resamp_UKB_global.Rmd <==
---
title: "Part 5-1: Resampling schemes that improve the effect sizes and power on global braim measures in UKB"
author: "Kaidi Kang"
date: "2023-08-14"
output: html_document
editor_options: 
  chunk_output_type: inline
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(cowplot) # for plot_grid
library(ggplot2)
library(pbmcapply)
library(magrittr)
library(dplyr)
library(readr)
library(splines)

```


```{r}
analysis_data <- read_csv("data/analysis_data/whole_brain_measure/new/analysis_data_whole_brain_combat.csv") 
analysis_data %<>% as.data.frame # `tibble` will cause a lot of problems with current RESI package
study_info_clean <- read_csv("data/analysis_data/whole_brain_measure/new/study_info_whole_brain.csv")
study_info_clean %<>% as.data.frame
```


# UKB

```{r, eval = TRUE}
# data UKB
UKB = subset(analysis_data, study == "UKB")

# winsoring
# ggplot(UKB, aes(x = age_years_winsor)) + geom_histogram()
UKB$age_years_winsor = UKB$age_years
UKB$age_years_winsor[UKB$age_years < 50] = 50
UKB$age_years_winsor[UKB$age_years > 78] = 78

Fn = ecdf(UKB$age_years_winsor)
UKB$freq = Fn(UKB$age_years_winsor + 0.5) - Fn(UKB$age_years_winsor - 0.5)

# Resampling probabilities
# s0: scheme 0 (flat)
UKB$resamp_prob_flat = 1/UKB$freq

x = UKB$age_years_winsor - mean(range(UKB$age_years_winsor))
# s1: scheme 1 (U shape)
UKB$resamp_prob_U =  (x^2 + 0.05 * max(x^2)) / (1.05*max(x^2 ))
# plot(UKB$age_years_winsor, UKB$resamp_prob_U)

# s2: scheme 2 (bell shape)
UKB$resamp_prob_inv_U = (- UKB$resamp_prob_U + 1.05)
# plot(UKB$age_years_winsor, UKB$resamp_prob_inv_U)

# data = UKB
# samp_size = 1000
# scheme = 2

resamp_func_ukb = function(data, samp_size, scheme){
  
  probs = data$resamp_prob_flat * data$resamp_prob_U^(scheme == 1) * data$resamp_prob_inv_U^(scheme == 2)
  # scheme 0
  if (scheme == 0) boot_id = sample(data$participant, size = samp_size, replace = TRUE, prob = probs)
  # scheme 1
  if (scheme == 1) boot_id = sample(data$participant, size = samp_size, replace = TRUE, prob = probs)
  # scheme 2
  if (scheme == 2) boot_id = sample(data$participant, size = samp_size, replace = TRUE, prob = probs)
  
  boot_data = data[unlist(lapply(boot_id, function(x) which(x == data[, "participant"]))), ]
  # table(boot_data$participant) %>% max()
  
  boot_val_sd = sd(boot_data$age_years)
  
  boot_lm = lm(GMV_10000_combat ~ ns(age_years, 2) + sex_01, data = boot_data)
  
  boot_resi = boot_lm %>% RESI::resi_pe(data = boot_data)
  # age effects
  boot_val_resi = boot_resi$anova[1, 4]
  # testing results for age effects
  boot_val_p = boot_resi$anova[1, 3]
  output = c(boot_val_sd, boot_val_resi, boot_val_p)
  names(output) = c("sd", "resi", "p_val")
  return(output)
}

```


```{r, eval = TRUE}
Sys.time()
set.seed(2023)
nboot = 1000
num.cores = 46
Ns = seq(20, 400, by = 5)
schemes = 0:2


output = NULL

for (N in Ns) {
  samp_size = N
  cat(samp_size, "\n")
  
  rv = data.frame(N = samp_size, scheme = 0:2, 
                  mean_RESI_age = NA, RESI_age_ll = NA, RESI_age_ul = NA, 
                  mean_sd_age = NA, sd_age_ll = NA, sd_age_ul = NA,
                  prob_larger_ES = NA, power = NA, replicability = NA)
  
  # scheme 0
  temp_0 = simplify2array(
            pbmclapply(1:nboot, function(simInd, data, samp_size, scheme){
              rep = resamp_func_ukb(data = data, samp_size = samp_size, scheme = scheme)
              return(rep)
            },
            data = UKB,
            samp_size = samp_size,
            scheme = 0,
            mc.cores = num.cores
            )) %>% t() %>% as.data.frame
  # scheme 1
  temp_1 = simplify2array(
          pbmclapply(1:nboot, function(simInd, data, samp_size, scheme){
            rep = resamp_func_ukb(data = data, samp_size = samp_size, scheme = scheme)
            return(rep)
          },
          data = UKB,
          samp_size = samp_size,
          scheme = 1,
          mc.cores = num.cores
          )) %>% t() %>% as.data.frame
  # scheme 2
  temp_2 = simplify2array(
          pbmclapply(1:nboot, function(simInd, data, samp_size, scheme){
            rep = resamp_func_ukb(data = data, samp_size = samp_size, scheme = scheme)
            return(rep)
          },
          data = UKB,
          samp_size = samp_size,
          scheme = 2,
          mc.cores = num.cores
          )) %>% t() %>% as.data.frame
  
  
  # Probability of observing a larger ES using a scheme than another
  ## scheme 1 vs 0 (U-shape vs flat)
  temp = expand.grid(s1_val = temp_1$resi, s0_val = temp_0$resi)
  rv[which(rv$scheme == 1), 'prob_larger_ES'] = mean(temp$s1_val >= temp$s0_val)
  ## scheme 2 vs 0 (inverse U-shape vs flat)
  temp = expand.grid(s2_val = temp_2$resi, s0_val = temp_0$resi)
  rv[which(rv$scheme == 2), 'prob_larger_ES'] = mean(temp$s2_val >= temp$s0_val)
  ## scheme 0 vs 0
  temp = expand.grid(s0_val_1 = temp_0$resi, s0_val = temp_0$resi)
  rv[which(rv$scheme == 0), 'prob_larger_ES'] = mean(temp$s0_val_1 >= temp$s0_val)
  
  # Power
  rv[which(rv$scheme == 0), 'power'] = mean(temp_0$p_val <= 0.05)
  rv[which(rv$scheme == 1), 'power'] = mean(temp_1$p_val <= 0.05)
  rv[which(rv$scheme == 2), 'power'] = mean(temp_2$p_val <= 0.05)

  # replicability
  ## def: the probability of observing the same results (both significant) from two random bootstraps
  # scheme 0 
  temp = expand.grid(boot_1 = temp_0$p_val, boot_2 = temp_0$p_val)
  rv[which(rv$scheme == 0), 'replicability'] = mean(temp$boot_1 <= 0.05 & temp$boot_2 <= 0.05)
  # scheme 1
  temp = expand.grid(boot_1 = temp_1$p_val, boot_2 = temp_1$p_val)
  rv[which(rv$scheme == 1), 'replicability'] = mean(temp$boot_1 <= 0.05 & temp$boot_2 <= 0.05)
  # scheme 0 
  temp = expand.grid(boot_1 = temp_2$p_val, boot_2 = temp_2$p_val)
  rv[which(rv$scheme == 2), 'replicability'] = mean(temp$boot_1 <= 0.05 & temp$boot_2 <= 0.05)
  
  # RESI 
  rv[which(rv$scheme == 0), 'mean_RESI_age'] = mean(temp_0$resi)
  rv[which(rv$scheme == 0), 'RESI_age_ll'] = quantile(temp_0$resi, probs = c(0.025))
  rv[which(rv$scheme == 0), 'RESI_age_ul'] = quantile(temp_0$resi, probs = c(0.975))
  
  rv[which(rv$scheme == 1), 'mean_RESI_age'] = mean(temp_1$resi)
  rv[which(rv$scheme == 1), 'RESI_age_ll'] = quantile(temp_1$resi, probs = c(0.025))
  rv[which(rv$scheme == 1), 'RESI_age_ul'] = quantile(temp_1$resi, probs = c(0.975))

  rv[which(rv$scheme == 2), 'mean_RESI_age'] = mean(temp_2$resi)
  rv[which(rv$scheme == 2), 'RESI_age_ll'] = quantile(temp_2$resi, probs = c(0.025))
  rv[which(rv$scheme == 2), 'RESI_age_ul'] = quantile(temp_2$resi, probs = c(0.975))
  
  # SD of age
  rv[which(rv$scheme == 0), 'mean_sd_age'] = mean(temp_0$sd)
  rv[which(rv$scheme == 0), 'sd_age_ll'] = quantile(temp_0$sd, probs = c(0.025))
  rv[which(rv$scheme == 0), 'sd_age_ul'] = quantile(temp_0$sd, probs = c(0.975))
  
  rv[which(rv$scheme == 1), 'mean_sd_age'] = mean(temp_1$sd)
  rv[which(rv$scheme == 1), 'sd_age_ll'] = quantile(temp_1$sd, probs = c(0.025))
  rv[which(rv$scheme == 1), 'sd_age_ul'] = quantile(temp_1$sd, probs = c(0.975))

  rv[which(rv$scheme == 2), 'mean_sd_age'] = mean(temp_2$sd)
  rv[which(rv$scheme == 2), 'sd_age_ll'] = quantile(temp_2$sd, probs = c(0.025))
  rv[which(rv$scheme == 2), 'sd_age_ul'] = quantile(temp_2$sd, probs = c(0.975))
  
  output = rbind(output, rv)
}

# output

saveRDS(output, paste0("Results/resamp_whole_brain_measures/resamp_scheme_UKB.rds"))
Sys.time()
```
