==> 5-2_Resamp_ADNI_global.Rmd <==
---
title: "Part 5-2: Longitudinal resampling schemes on global brain measures in ADNI"
author: "Kaidi Kang"
date: "2023-08-23"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


```{r}
library(readr)
library(Hmisc) # for binomial wilson confidence interval
library(cowplot) # for plot_grid & get_legend
library(pbmcapply)
library(magrittr)
library(dplyr)
library(ggplot2)
library(RESI)
library(geepack)
library(splines)
```



# ADNI
New goal: ACTIONABLE results. 

Make (1) the baseline covariate and (2) the change in covariate (here, age in years) since baseline follow specific distributions, respectively. 

## Sampling weights for follow-up and the baseline 
```{r}
analysis_data <- read_csv("RESI/longitudinal/data_analysis/1st_revision/data/analysis_data/whole_brain_measure/new/analysis_data_whole_brain_combat.csv")
analysis_data %<>% as.data.frame # `tibble` will cause a lot of problems with current RESI package

ADNI = subset(analysis_data, study == "ADNI")

ADNI %<>% arrange(participant, age_years)

ADNI %<>% group_by(participant) %>% mutate(age_bl = min(age_years),
                                           D_age = age_years - age_bl,
                                           visit = row_number(),
                                           num_visit = n())
```


```{r}
# hist of baseline age in ADNI
ggplot(data = subset(ADNI, visit == 1), aes(x = age_bl)) + geom_histogram()
# hist of D in ADNI
ggplot(data = subset(ADNI, visit != 1), aes(x = D_age)) + geom_histogram()
# their correlation
ggplot(data = subset(ADNI, visit != 1), aes(x = age_bl, y = D_age)) + geom_point()
```


```{r}
ADNI_bl = subset(ADNI, visit == 1)
# ADNI_fu = subset(ADNI, visit != 1)

var_list = c("participant", "GMV_10000_combat", "age_years",  "sex_01", "visit")

ADNI_fu = merge(ADNI_bl[var_list], ADNI[var_list], by = "participant", all = TRUE)

ADNI_fu %<>% rename(GMV_10000_combat_0 = GMV_10000_combat.x, 
                        age_years_0 = age_years.x,
                        visit_0 = visit.x,
                        sex_01 = sex_01.x,
                        GMV_10000_combat_1 = GMV_10000_combat.y, 
                        age_years_1 = age_years.y,
                        visit_1 = visit.y)

# calculate D(age)
ADNI_fu$D_age = abs(ADNI_fu$age_years_1 - ADNI_fu$age_years_0)

# winsorizing
ADNI_fu$age_years_0_winsor = ADNI_fu$age_years_0
ADNI_fu$age_years_0_winsor = ifelse(ADNI_fu$age_years_0 >= 85, 85, ADNI_fu$age_years_0_winsor)
ADNI_fu$age_years_0_winsor = ifelse(ADNI_fu$age_years_0_winsor <= 65, 65, ADNI_fu$age_years_0_winsor)
ADNI_fu$D_age_winsor = ADNI_fu$D_age
ADNI_fu$D_age_winsor = ifelse(ADNI_fu$D_age >= 5, 5, ADNI_fu$D_age)

# derive the empirical bivariate density 

ADNI_fu$freq = lapply(1:nrow(ADNI_fu), FUN = function(x){
  a = ADNI_fu$age_years_0_winsor[x]
  b = ADNI_fu$D_age_winsor[x]
  d = sum(ADNI_fu$age_years_0_winsor >= a - 0.5 & ADNI_fu$age_years_0_winsor <= a + 0.5 & ADNI_fu$D_age_winsor >= b - 0.5 & ADNI_fu$D_age_winsor <= b + 0.5) / nrow(ADNI_fu)
  return(d)
}) %>% unlist

# weight for uniform resampling
ADNI_fu$resamp_prob_flat = 1 / ADNI_fu$freq

# U shape
x = ADNI_fu$age_years_0_winsor - mean(range(ADNI_fu$age_years_0_winsor))
ADNI_fu$bl_resamp_prob_U =  (x^2 + 0.05 * max(x^2) ) / (1.05*max(x^2 ))
# bell shape
ADNI_fu$bl_resamp_prob_bell = - ADNI_fu$bl_resamp_prob_U + 1.05

# Target dist for change in age
# increasing shape
ADNI_fu$D_resamp_prob_inc = (exp(ADNI_fu$D_age_winsor) + 0.05 * exp(5) )/ max(exp(ADNI_fu$D_age_winsor) + 0.05 * exp(5) )

# decreasing shape
ADNI_fu$D_resamp_prob_dec = (exp(-ADNI_fu$D_age_winsor) + 0.05 )/ max(exp(-ADNI_fu$D_age_winsor) + 0.05  )
```


## Bootstraping (m = 2)

```{r}
# # bl_scheme: 0 - uniform; 1 - U-shape; 2 - Bell-shape
# # D_scheme: 0 - uniform; 1 - Increasing; 2 - Decreasing

resamp_ADNI_2obs = function(data, samp_size, bl_scheme, D_scheme){
  
  probs = data$resamp_prob_flat * (data$bl_resamp_prob_U)^(bl_scheme == 1) * (data$bl_resamp_prob_bell)^(bl_scheme == 2) * (data$D_resamp_prob_inc)^(D_scheme == 1) * (data$D_resamp_prob_dec)^(D_scheme == 2)
   
  bad = TRUE
   
  err_msg = NA

  while (bad){
  
    boot_id = sample(1:nrow(data), size = samp_size, replace = TRUE, prob = probs )
    boot_data = lapply(1:length(boot_id), function(x) {
      id = boot_id[x]
      select_dat = data[id, ]
      v1 = select_dat[, c("participant", "GMV_10000_combat_0", "age_years_0", "sex_01", "D_age", "visit_0")]
      v2 = select_dat[, c("participant", "GMV_10000_combat_1", "age_years_1", "sex_01", "D_age", "visit_1")]
      v1 %<>% rename(GMV_10000_combat = GMV_10000_combat_0,
                     age_years = age_years_0,
                     visit = visit_0)
      v2 %<>% rename(GMV_10000_combat = GMV_10000_combat_1,
                     age_years = age_years_1,
                     visit = visit_1)
      subj_dat = rbind(v1[, c("participant", "GMV_10000_combat", "age_years", "sex_01", "D_age", "visit")],
                       v2[, c("participant", "GMV_10000_combat", "age_years", "sex_01", "D_age", "visit")])
      subj_dat$temp_id = x 
      subj_dat$new_visit = 1:2
      return(subj_dat)
    })
    boot_data = do.call(rbind, boot_data)
    boot_data %<>% arrange(temp_id, age_years)
  
    # ANALYSIS
    ## mean of baseline age
    mean_bl_age = boot_data$age_years[boot_data$visit == 1] %>% mean()
    ## SD of baseline age
    sd_bl_age = boot_data$age_years[boot_data$visit == 1] %>% sd()
    ## mean of D_age
    mean_D_age = boot_data$D_age[boot_data$visit != 1] %>% mean()
    ## SD of D_age
    sd_D_age = boot_data$D_age[boot_data$visit != 1] %>% sd()
  
    # calculate the between- & within-subject SD of age in the bootstrap data 
    # using a linear mixed effects model
    lme_fit = lmer(age_years ~ (1 | temp_id), data = boot_data)
    ran_eff_var = VarCorr(lme_fit) %>% as.data.frame
    # between-subject mean age
    btwn_subj_mean_age = summary(lme_fit)$coefficients[1]
    # between-subject sd of age
    btwn_subj_sd_age = ran_eff_var$vcov[1] %>% sqrt()
    # within-subject sd of age
    within_subj_sd_age = ran_eff_var$vcov[2] %>% sqrt()
    
    # ES of age
    boot_gee  =  tryCatch({
        geeglm(GMV_10000_combat ~ sex_01 + ns(age_years, 2), data = boot_data, corstr = "exchangeable", id = temp_id)
      },
      error = function(e){
        bad <<- TRUE
        err_msg = e
      })
    
    if (is.na(err_msg)) bad <- FALSE
    
  } # while loop

  # RESI
  boot_resi = boot_gee %>% RESI::resi_pe(data = boot_data)

  boot_val_CS_resi = boot_resi$anova[2, 'CS-RESI'] # this is type 1 sequential anova
  boot_val_L_resi = boot_resi$anova[2, 'L-RESI']
  
  # p-val for age effects
  boot_val_p = boot_resi$anova[2, 3]
  
  output = c(mean_bl_age, sd_bl_age, mean_D_age, sd_D_age, btwn_subj_sd_age, within_subj_sd_age, boot_val_CS_resi, boot_val_L_resi, boot_val_p, err_msg)
                 
  names(output) = c("mean_bl_age", "sd_bl_age", "mean_D_age", "sd_D_age", "btwn_subj_sd_age", "within_subj_sd_age", "CS_RESI", "L_RESI", "p_val", "err_msg")
   
  return(output)
}

```

```{r}
Sys.time()
set.seed(2023)
Ns = seq(20, 150, by = 5)
nboot = 1000
num_cores = 46
# between-subject resampling scheme.
bl_schemes = 0:2
D_schemes = 0:2


output = expand.grid(N = Ns, nboot = nboot, bl_scheme = bl_schemes, D_scheme = D_schemes)


for (i in 1:nrow(output)){

  samp_size = output$N[i]
  bl_s = output$bl_scheme[i]
  D_s = output$D_scheme[i]
  
  cat(i, "- N:", samp_size, "  - bl scheme = ", bl_s, "  - D scheme = ", D_s, "\n")

  temp <- lapply(1:nboot, function(simInd, data, samp_size, bl_scheme, D_scheme){
            rep = resamp_ADNI_2obs(data = data, samp_size = samp_size, bl_scheme = bl_scheme, D_scheme = D_scheme)
            return(rep)
          },
          data = ADNI_fu,
          samp_size = samp_size,
          bl_scheme = bl_s,
          D_scheme = D_s
          # ,
          # mc.cores = num_cores
          )
  temp = do.call(rbind, temp) %>% as.data.frame
  
  print(unique(temp$err_msg))
  
  # mean
  output[which(output$N == samp_size & output$bl_scheme == bl_s & output$D_scheme == D_s), 
         c("bl_age", "sd_bl_age", "D_age", "sd_D_age", "btwn_subj_sd_age", "within_subj_sd_age", "CS_RESI_age", "L_RESI_age")] = colMeans(temp[, 1:8])

  # lower bound
  output[which(output$N == samp_size & output$bl_scheme == bl_s & output$D_scheme == D_s), 
         c("bl_age_ll", "sd_bl_age_ll", "D_age_ll", "btwn_subj_sd_age_ll", "within_subj_sd_age_ll", "sd_D_age_ll", "CS_RESI_age_ll", "L_RESI_age_ll")] =  apply(temp[, 1:8], 2, quantile, probs = c(0.025))
  
  # upper bound
  output[which(output$N == samp_size & output$bl_scheme == bl_s & output$D_scheme == D_s), 
         c("bl_age_ul", "sd_bl_age_ul", "D_age_ul", "btwn_subj_sd_age_ul", "within_subj_sd_age_ul", "sd_D_age_ul", "CS_RESI_age_ul", "L_RESI_age_ul")] =  apply(temp[, 1:8], 2, quantile, probs = c(0.975))
  
  # Power
  output[which(output$N == samp_size & output$bl_scheme == bl_s & output$D_scheme == D_s), 'power'] = mean(temp$p_val <= 0.05)

  # replicability
  ## def: the probability of observing the same results (both significant) from two random bootstraps
  # scheme 0 
  grid = expand.grid(boot_1 = temp$p_val, boot_2 = temp$p_val)
  output[which(output$N == samp_size & output$bl_scheme == bl_s & output$D_scheme == D_s), 'replicability'] = mean(grid$boot_1 <= 0.05 & grid$boot_2 <= 0.05)
  
}
Sys.time()
saveRDS(output, paste0("Results/resamp_whole_brain_measures/resamp_scheme_ADNI_m=2.rds"))

```
